esap/requirements/base.txt
Django==3.1.4
django-filter==2.3.0
djangorestframework==3.11.1
django-cors-headers==3.7.0
# cx-Oracle==7.2.2
whitenoise==5.0.1
numpy==1.18.1
astropy==4.0
pyvo==1.0
requests==2.22.0
django-rest-knox==4.1.0
mozilla-django-oidc==1.2.4
panoptes-client==1.2.0
